+++ CHANGELOG.md
@@ -11,6 +11,24 @@ and this project adheres to
 
 ### Added
 
+- Added support to publish job log events at level `info`, `warn`, and `error`.
+
+  Usage:
+
+  ```ts
+  logger.publishInfoEvent({
+    name: IntegrationErrorEventName.Stats,
+    description: 'fetched 100000 records',
+  });
+  logger.publishWarnEvent({
+    name: IntegrationErrorEventName.MissingPermission,
+    description: 'Missing permission users.read',
+  });
+  logger.publishErrorEvent({
+    name: IntegrationErrorEventName.MissingPermission,
+    description: 'Missing permission users.read',
+  });
+  ```
 - Added support for relative paths in `yarn j1-integration *` commands
 
 ## [7.0.0] - 2021-10-05